코드 리팩토링

+++ BookShop/controller/CartController.js
@@ -1,24 +1,34 @@
 const jwt = require('jsonwebtoken')
 const conn = require('../mariadb');
-const {StatusCodes} = require('http-status-codes');
-const dotenv = require('dotenv');
-dotenv.config()
+const { StatusCodes } = require('http-status-codes');
+const ensureAuthorization = require('../auth');
 
 
 const addToCart = (req, res) => {
-    const {book_id, quantity} = req.body;
+    const { book_id, quantity } = req.body;
     let authorization = ensureAuthorization(req)
 
-    let sql = "INSERT INTO cartItems (book_id, quantity, user_id) VALUES (?, ?, ?);"
-    let values = [book_id, quantity, authorization.id]
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        });
+    } else {
+        let sql = "INSERT INTO cartItems (book_id, quantity, user_id) VALUES (?, ?, ?);"
+        let values = [book_id, quantity, authorization.id]
 
-    conn.query(sql, values, (err, results) => {
-        if (err) {
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+        conn.query(sql, values, (err, results) => {
+            if (err) {
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
 
-        return res.status(StatusCodes.OK).json(results);
-    })
+            return res.status(StatusCodes.OK).json(results);
+        })
+    }
+
 };
 
 const getCartItems = (req, res) => {
@@ -27,44 +37,70 @@ const getCartItems = (req, res) => {
 
     let authorization = ensureAuthorization(req)
 
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        });
+    } else {
 
-    let sql = `SELECT cartitems.id, book_id, title, summary, quantity, price
+        let sql = `SELECT cartitems.id, book_id, title, summary, quantity, price
                 FROM cartitems LEFT JOIN books
                 ON cartitems.id = books.id
-                WHERE user_id = ? AND id IN (?)`
+                WHERE user_id = ?`
 
-    const values = [authorization.id, selected]
+        const values = [authorization.id]
 
-    conn.query(sql, values, (err, results) => {
-        if (err) {
-            return res.status(StatusCodes.BAD_REQUEST).end();
+
+        if (selected) {
+            sql += ` AND cartItems.id IN (?)`
+            values.push(selected)
         }
 
-        return res.status(StatusCodes.OK).json(results);
-    })
+
+        conn.query(sql, values, (err, results) => {
+            if (err) {
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+
+            return res.status(StatusCodes.OK).json(results);
+        })
+    }
+
 };
 
 const removeCartItem = (req, res) => {
-    const cartItemId = req.params.id;
+    let authorization = ensureAuthorization(req)
 
-    let sql = "DELETE FROM cartitems WHERE id = ?"
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        });
+    } else {
+        const cartItemId = req.params.id;
 
-    conn.query(sql, cartItemId, (err, results) => {
-        if (err) {
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+        let sql = "DELETE FROM cartitems WHERE id = ?"
 
-        return res.status(StatusCodes.OK).json(results);
-    })
+        conn.query(sql, cartItemId, (err, results) => {
+            if (err) {
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+
+            return res.status(StatusCodes.OK).json(results);
+        })
+    }
+
+
+
 };
 
-function ensureAuthorization(req) {
-    let receivedJwt = req.headers["authorization"];
-
-    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-
-    return decodedJwt
-}
 
 module.exports = {
     addToCart,
